utils: Move behind feature flag
There was a circular dependency between the opentok crate and the utils crate...
As we don't really need opentok to depend on GStreamer, move these dependencies
behind a `utils` feature, which is then needed to run the tests and examples.
