+++ .github/workflows/rust.yml
@@ -29,7 +29,7 @@ jobs:
     - name: Build
       run: cargo build --verbose
     - name: Run tests
-      run: cargo test --verbose -- --test-threads=1
+      run: cargo test --features gstreamer-utils --verbose -- --test-threads=1
       env:
         OPENTOK_KEY: ${{ secrets.OPENTOK_KEY }}
         OPENTOK_SECRET: ${{ secrets.OPENTOK_SECRET }}
