+++ examples/src/bin/subscriber_only.rs
@@ -1,7 +1,7 @@
 extern crate opentok;
 
 use opentok::log::{self, LogLevel};
-use opentok_utils::subscriber::Subscriber;
+use opentok::utils::subscriber::Subscriber;
 
 #[path = "../cli.rs"]
 mod cli;
