--- utils/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "opentok-utils"
-version = "1.0.1"
-edition = "2021"
-description = "Test utilities for the OpenTok SDK Rust bindings"
-authors = ["Fernando Jiménez Moreno <ferjmoreno@gmail.com>", "Philippe Normand <philn@igalia.com>"]
-repository = "https://github.com/opentok-rust/opentok-rs"
-documentation = "https://opentok-rust.github.io/opentok-rs/opentok_utils/"
-license = "MIT"
-
-[dependencies]
-anyhow = "1.0.40"
-byte-slice-cast = "1.0.0"
-derive_more = "0.99.14"
-glib = "0.15"
-gst = { package = "gstreamer", version = "0.18" }
-gst_app = { package = "gstreamer-app", version = "0.18" }
-gst_video = { package = "gstreamer-video", version = "0.18" }
-opentok = { path = "../opentok" }
